+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.4] - Unreleased
+
+### Fixed
+
+- [#188] Skip borrower renewal API call when due date with buffer is in the past
+- [#188] Treat Rapido "dueDate already passed" 500 responses as permanent task failure
+
 ## [1.5.2] - 2026-04-10
 
 ### Fixed